+++ modules-template/prov-template/src/main/java/org/openprovenance/prov/template/expander/ttf/ConfigTask.java
@@ -1,4 +1,4 @@
-package org.openprovenance.prov.template.expander.meta;
+package org.openprovenance.prov.template.expander.ttf;
 
 import java.io.IOException;
 import java.util.Map;