+++ modules-template/prov-template/src/main/java/org/openprovenance/prov/template/expander/ttf/MergeTask.java
@@ -1,4 +1,4 @@
-package org.openprovenance.prov.template.expander.meta;
+package org.openprovenance.prov.template.expander.ttf;
 
 import org.apache.commons.lang3.tuple.Pair;
 import org.openprovenance.prov.model.Document;