+++ RELEASES.md
@@ -6,6 +6,7 @@ See also GitHub releases https://github.com/lucmoreau/ProvToolbox/releases
 * prov-template: updated BatchExecutor to create provenance when merging templates
 * prov-template: refactored BatchExecutor to support two different tasks for instantiation and merge
 * prov-template: merge tasks now supporting a list of input files
+* prov-template: renamed ttb to ttf as per book
  
 `2.2.1`
 * service-metrics: a service to gather metrics on provenance, using MetricsCalculator.java